Fix browser CI: move browser-presence guard out of LeakCanary harness; drop <queries>
The browserIsInstalledOnDevice guard inherited BasePlaygroundTest's
DetectLeaksAfterTestSuccess rule, so its trivial success triggered a heap check that
pinned a pre-existing app leak on it (NoLeakAssertionFailedError) and, via retry x3,
ballooned shard _0 to ~13min. Move it to a standalone BrowserAvailabilityTest that
doesn't extend BasePlaygroundTest (no leak detection / playground setup).

Also remove the androidTest <queries> block: the app-under-test already exposes
browsers to getInstalledApplications() via androidx.browser's Custom Tabs queries, so
it's likely redundant. The standalone guard now fails loudly (clean assertion) if a
browser is genuinely not visible, so removing it is safe to verify on CI.

Committed-By-Agent: claude
