+++ paymentsheet-example/src/androidTest/java/com/stripe/android/TestBrowsers.kt
@@ -1,12 +1,8 @@
 package com.stripe.android
 
-import android.content.pm.PackageManager
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.test.platform.app.InstrumentationRegistry
-import com.google.common.truth.Truth.assertThat
 import com.stripe.android.test.core.Browser
 import com.stripe.android.test.core.TestParameters
-import com.stripe.android.test.core.ui.BrowserUI
 import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -25,21 +21,6 @@ internal class TestBrowsers : BasePlaygroundTest() {
     ).copy(
         useBrowser = Browser.Chrome,
     )
-
-    /**
-     * Guards the browser device: browser-redirect tests Assume-skip when no browser is detected, so
-     * a missing/invisible browser would silently drop their coverage. This fails loudly instead.
-     */
-    @Test
-    fun browserIsInstalledOnDevice() {
-        val knownBrowserPackages = BrowserUI.values().map { it.packageName }
-        val installedBrowsers = InstrumentationRegistry.getInstrumentation().targetContext
-            .packageManager
-            .getInstalledApplications(PackageManager.GET_META_DATA)
-            .map { it.packageName }
-            .filter { it in knownBrowserPackages }
-        assertThat(installedBrowsers).isNotEmpty()
-    }
 
     @Test
     fun testAuthorizeChrome() {
